requirements.txt
requests~=2.32.5
flask~=3.1.2
langchain~=0.3.27
langgraph~=0.6.7
pypdf~=6.0.0
langchain-text-splitters~=0.3.11
sentence-transformers~=5.1.0
langchain-core~=0.3.76
langchain-community~=0.3.29
python-dotenv~=1.1.1